Refactor p1.f for enhanced readability
Refactored the p1.f to use functions for enhanced segmentation and
readability. This allows for easier debugging and reduced likelihood of
side effects and regression.

In addition, the multiple do loops have been merged to reduce verbosity
of implementation. It was unnecessary to have four do loops in the
initial implementation as there's an order of operations. This naivete
will be corrected in future commits.