Change show_staticness() to integer return type
Changed the show_staticness() function to integer return type. The
assignment instructions specify an integer function for demonstrating
staticness. The change is introducing show_staticness() as an integer
variable within the program/end block. Otherwise, the program assumes
the return type is real and reports an error when compiled with
gfortran.